+++ hibernate-core/src/main/java/org/hibernate/type/jakartajson/JsonBJsonFormatMapper.java
@@ -4,8 +4,9 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.type;
+package org.hibernate.type.jakartajson;
 
+import org.hibernate.type.FormatMapper;
 import org.hibernate.type.descriptor.WrapperOptions;
 import org.hibernate.type.descriptor.java.JavaType;
 
@@ -16,10 +17,9 @@ import jakarta.json.bind.JsonbException;
 /**
  * @author Christian Beikov
  */
-public class JsonBJsonFormatMapper implements FormatMapper {
+public final class JsonBJsonFormatMapper implements FormatMapper {
 
 	public static final String SHORT_NAME = "jsonb";
-	public static final JsonBJsonFormatMapper INSTANCE = new JsonBJsonFormatMapper();
 
 	private final Jsonb jsonb;
